Add Proposed Schema
I put together a document that uses a [Mermaid Entity Relationship
Diagram](https://mermaid-js.github.io/mermaid/#/entityRelationshipDiagram)
to put what you sent in the email together with what I had in my head.

Once we agree on what we should store in the database, I think we should
merge this and then from that we'd create the actual Postgres schema.