Setup the repo for Flask
This change does a bunch of stuff to get us initially setup.

1. It sets up the `trv` package, which is under the `src/` directory,
   and adds the `setup.py` file to produce an installable package.
2. Inside the `trv` package it does some boilerplate stuff in
   `src/trv/__init__.py` to create the Flask app, and then creates one
Blueprint for an `/api` set of routes. Those are the "backend" routes
we'll have for updating or working with the database. We'll probably
have another Blueprint maybe called `views` or something that will
render the HTML pages, like an `events.html`.
3. It sets up a `tests/` directory that has some tests of the routes
   that this change adds.
4. This sets up the requirements we'll need and does it via a similar
   tactic to what I use at work. We'll want `requirements.txt` to not
move unless we tell it to, so we use this concept of putting loose
requirements in `base-requirements.txt` and `requirements.txt` becomes
the "baked" version of it via the `pip freeze` command.
5. Adds a little `scripts/flask-dev.py` script so you can run the dev
   server without having to deal with environment variables or anything
like that.
6. This sets up `pre-commit` so our code is formatted in a uniform way,
   using a tool called `black` along with doing some things like always
trimming whitespace from the ends of lines. When working with multiple
people, I've found these sorts of tools really helpful because it takes
away a lot of needless differences and keeps those out of code reviews.
When we let a tool do the formatting, personal preference for how things
look becomes a lot less of what matters in a code review, so we can keep
the reviews more focused.